Add useLatestStudyDetail hook
